Increase the size of enum discriminant

+++ casper-contract-schema/src/lib.rs
@@ -174,7 +174,7 @@ pub enum CustomType {
 pub struct EnumVariant {
     pub name: String,
     pub description: Option<String>,
-    pub discriminant: u8,
+    pub discriminant: u16,
     pub ty: Type,
 }
